+++ collector-bitmovin-player/src/androidTest/java/com/bitmovin/analytics/bitmovin/player/BitmovinPlayerConstants.kt
@@ -3,5 +3,5 @@ package com.bitmovin.analytics.bitmovin.player
 import com.bitmovin.analytics.test.utils.PlayerInfo
 
 object BitmovinPlayerConstants {
-    val playerInfo = PlayerInfo("bitmovin", "bitmovin-3.138.0", "Android:Exoplayer")
+    val playerInfo = PlayerInfo("bitmovin", "bitmovin-3.154.0", "Android:Exoplayer")
 }